feat: Add Login component
The App will return the Login component if
AuthenticationContext.authenticated is falsey.  Then the
AuthenticationProvider can be used to provide an async logic function
to attempt authentication.

If the user is authenticated, the token is added to sessionStorage (so
that it can be used to make authenticated API calls), and the user
is pushed to whichever route they were attempting to reach.